extra: fonts: Add Apple SD Gothic Neo v19.0d2e1
Apple product users are allowed to use this font, but redistribution is
not allowed. As per Apple, You may not rent, lease, lend, trade,
transfer, sell, sublicense or otherwise redistribute the Apple Font in
any unauthorized way[0]. Additionally, this font is not a VF, and its
file size is relatively large. Despite this, I have yet to find another
font that matches the UI as well as this one does.

The font file was obtained from from macOS 23E214 and extracted with
FontForge. It was then converted back and forth without any modification
using the ttx tool to fix general errors.

[0]: https://developer.apple.com/fonts/